chore(dyn-abi): gate eip712 behind a feature
