+++ crates/dyn-abi/src/error.rs
@@ -1,4 +1,3 @@
-use crate::DynSolType;
 use alloc::string::{String, ToString};
 use core::fmt;
 
@@ -10,17 +9,19 @@ pub type DynAbiResult<T, E = DynAbiError> = core::result::Result<T, E>;
 /// <https://eips.ethereum.org/EIPS/eip-712#definition-of-encodetype>
 #[derive(Debug, Clone, PartialEq)]
 pub enum DynAbiError {
-    /// Type mismatch during coercion.
-    TypeMismatch {
-        /// The expected type.
-        expected: DynSolType,
-        /// The actual type.
-        actual: serde_json::Value,
-    },
     /// Invalid size for a primitive type (intX, uintX, or bytesX).
     InvalidSize(String),
     /// Invalid type string, extra chars, or invalid structure.
     InvalidTypeString(String),
+
+    /// Type mismatch during coercion.
+    #[cfg(feature = "eip712")]
+    TypeMismatch {
+        /// The expected type.
+        expected: crate::DynSolType,
+        /// The actual type.
+        actual: serde_json::Value,
+    },
     /// Unknown type referenced from another type.
     MissingType(String),
     /// Detected circular dep during typegraph resolution.
@@ -45,16 +46,19 @@ impl std::error::Error for DynAbiError {
 impl fmt::Display for DynAbiError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            DynAbiError::InvalidSize(ty) => write!(f, "Invalid size for type: {ty}"),
+            DynAbiError::InvalidTypeString(ty) => write!(f, "Invalid type string: {ty}"),
+
+            #[cfg(feature = "eip712")]
             DynAbiError::TypeMismatch { expected, actual } => {
                 write!(f, "Type mismatch, expected: {expected:?}, actual: {actual}")
             }
-            DynAbiError::InvalidSize(ty) => write!(f, "Invalid size for type: {ty}"),
-            DynAbiError::InvalidTypeString(ty) => write!(f, "Invalid type string: {ty}"),
             DynAbiError::MissingType(name) => write!(f, "Missing type in type resolution: {name}"),
             DynAbiError::CircularDependency(dep) => write!(f, "Circular dependency: {dep}"),
             DynAbiError::InvalidPropertyDefinition(def) => {
                 write!(f, "Invalid property definition: {def}")
             }
+
             DynAbiError::HexError(h) => h.fmt(f),
         }
     }
@@ -66,8 +70,21 @@ impl From<hex::FromHexError> for DynAbiError {
     }
 }
 
+#[allow(dead_code)]
 impl DynAbiError {
-    pub(crate) fn type_mismatch(expected: DynSolType, actual: &serde_json::Value) -> DynAbiError {
+    pub(crate) fn invalid_size(ty: impl ToString) -> DynAbiError {
+        DynAbiError::InvalidSize(ty.to_string())
+    }
+
+    pub(crate) fn invalid_type_string(ty: impl ToString) -> DynAbiError {
+        DynAbiError::InvalidTypeString(ty.to_string())
+    }
+
+    #[cfg(feature = "eip712")]
+    pub(crate) fn type_mismatch(
+        expected: crate::DynSolType,
+        actual: &serde_json::Value,
+    ) -> DynAbiError {
         DynAbiError::TypeMismatch {
             expected,
             actual: actual.clone(),
@@ -76,14 +93,6 @@ impl DynAbiError {
 
     pub(crate) fn invalid_property_def(def: impl ToString) -> DynAbiError {
         DynAbiError::InvalidPropertyDefinition(def.to_string())
-    }
-
-    pub(crate) fn invalid_size(ty: impl ToString) -> DynAbiError {
-        DynAbiError::InvalidSize(ty.to_string())
-    }
-
-    pub(crate) fn invalid_type_string(ty: impl ToString) -> DynAbiError {
-        DynAbiError::InvalidTypeString(ty.to_string())
     }
 
     pub(crate) fn missing_type(name: impl ToString) -> DynAbiError {
