+++ crates/dyn-abi/src/type.rs
@@ -1,7 +1,5 @@
 use crate::{
-    eip712::coerce::{self, coerce_custom_struct, coerce_custom_value},
-    parser::TypeSpecifier,
-    DynAbiError, DynAbiResult, DynSolValue, DynToken, Result, SolType, Word,
+    parser::TypeSpecifier, DynAbiError, DynAbiResult, DynSolValue, DynToken, Result, SolType, Word,
 };
 use alloc::{borrow::Cow, boxed::Box, string::String, vec::Vec};
 use alloy_sol_types::sol_data;
@@ -304,28 +302,6 @@ impl DynSolType {
         }
     }
 
-    /// Coerce a json value to a sol value via this type.
-    pub fn coerce(&self, value: &serde_json::Value) -> Result<DynSolValue, DynAbiError> {
-        match self {
-            DynSolType::Address => coerce::address(value),
-            DynSolType::Bytes => coerce::bytes(value),
-            DynSolType::Int(n) => coerce::int(*n, value),
-            DynSolType::Uint(n) => coerce::uint(*n, value),
-            DynSolType::Bool => coerce::bool(value),
-            DynSolType::Array(inner) => coerce::array(inner, value),
-            DynSolType::String => coerce::string(value),
-            DynSolType::FixedBytes(n) => coerce::fixed_bytes(*n, value),
-            DynSolType::FixedArray(inner, n) => coerce::fixed_array(inner, *n, value),
-            DynSolType::Tuple(inner) => coerce::tuple(inner, value),
-            DynSolType::CustomStruct {
-                name,
-                prop_names,
-                tuple,
-            } => coerce_custom_struct(name, prop_names, tuple, value),
-            DynSolType::CustomValue { name } => coerce_custom_value(name, value),
-        }
-    }
-
     /// Instantiate an empty dyn token, to be decoded into.
     pub(crate) fn empty_dyn_token(&self) -> DynToken<'_> {
         match self {
@@ -409,9 +385,7 @@ impl DynSolType {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloc::{borrow::ToOwned, string::ToString};
     use alloy_primitives::{hex, Address};
-    use serde_json::json;
 
     #[test]
     fn dynamically_encodes() {
@@ -439,113 +413,6 @@ mod tests {
         let mut enc = crate::Encoder::default();
         DynSolValue::encode_sequence(val.as_fixed_seq().unwrap(), &mut enc);
         assert_eq!(enc.finish(), vec![word1, word2]);
-    }
-
-    #[test]
-    fn it_coerces() {
-        let j = json!({
-            "message": {
-                "contents": "Hello, Bob!",
-                "attachedMoneyInEth": 4.2,
-                "from": {
-                    "name": "Cow",
-                    "wallets": [
-                        "0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826",
-                        "0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF",
-                    ]
-                },
-                "to": [{
-                    "name": "Bob",
-                    "wallets": [
-                        "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB",
-                        "0xB0BdaBea57B0BDABeA57b0bdABEA57b0BDabEa57",
-                        "0xB0B0b0b0b0b0B000000000000000000000000000",
-                    ]
-                }]
-            }
-        });
-
-        let ty = DynSolType::CustomStruct {
-            name: "Message".to_owned(),
-            prop_names: vec!["contents".to_string(), "from".to_string(), "to".to_string()],
-            tuple: vec![
-                DynSolType::String,
-                DynSolType::CustomStruct {
-                    name: "Person".to_owned(),
-                    prop_names: vec!["name".to_string(), "wallets".to_string()],
-                    tuple: vec![
-                        DynSolType::String,
-                        DynSolType::Array(Box::new(DynSolType::Address)),
-                    ],
-                },
-                DynSolType::Array(Box::new(DynSolType::CustomStruct {
-                    name: "Person".to_owned(),
-                    prop_names: vec!["name".to_string(), "wallets".to_string()],
-                    tuple: vec![
-                        DynSolType::String,
-                        DynSolType::Array(Box::new(DynSolType::Address)),
-                    ],
-                })),
-            ],
-        };
-        let top = j.as_object().unwrap().get("message").unwrap();
-
-        assert_eq!(
-            ty.coerce(top).unwrap(),
-            DynSolValue::CustomStruct {
-                name: "Message".to_owned(),
-                prop_names: vec!["contents".to_string(), "from".to_string(), "to".to_string()],
-                tuple: vec![
-                    DynSolValue::String("Hello, Bob!".to_string()),
-                    DynSolValue::CustomStruct {
-                        name: "Person".to_owned(),
-                        prop_names: vec!["name".to_string(), "wallets".to_string()],
-                        tuple: vec![
-                            DynSolValue::String("Cow".to_string()),
-                            vec![
-                                DynSolValue::Address(
-                                    "0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826"
-                                        .parse()
-                                        .unwrap()
-                                ),
-                                DynSolValue::Address(
-                                    "0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF"
-                                        .parse()
-                                        .unwrap()
-                                ),
-                            ]
-                            .into()
-                        ]
-                    },
-                    vec![DynSolValue::CustomStruct {
-                        name: "Person".to_owned(),
-                        prop_names: vec!["name".to_string(), "wallets".to_string()],
-                        tuple: vec![
-                            DynSolValue::String("Bob".to_string()),
-                            vec![
-                                DynSolValue::Address(
-                                    "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB"
-                                        .parse()
-                                        .unwrap()
-                                ),
-                                DynSolValue::Address(
-                                    "0xB0BdaBea57B0BDABeA57b0bdABEA57b0BDabEa57"
-                                        .parse()
-                                        .unwrap()
-                                ),
-                                DynSolValue::Address(
-                                    "0xB0B0b0b0b0b0B000000000000000000000000000"
-                                        .parse()
-                                        .unwrap()
-                                ),
-                            ]
-                            .into()
-                        ]
-                    }]
-                    .into()
-                ]
-            }
-        )
     }
 
     // also tests the type name parser
